moved SubNav to AllStudiesPage folder
