+++ src/pages/AllStudiesPage/AllStudiesPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import SideBar from '../../default/SideBar';
-import SubNav from '../../default/SubNav';
+import SubNav from './components/SubNav';
 import SearchBar from './components/SearchBar';
 import StudyList from './components/StudyList';
 import StudyStatus from './components/StudyStatus';
