+++ src/pages/AllStudiesPage/components/SearchBar.css
@@ -1,13 +1,14 @@
 .study-search {
   position: relative;
-  width: 100%;
+  width: 98%;
   margin-bottom: 1.5em;
+  /* padding-right: 1em; */
 }
 
 .study-search input {
-  width: 100%;
-  padding: 0.7rem;
-  padding-right: 0.2rem;
+  width: 93%;
+  padding: 0.85rem;
+  padding-right: 2rem;
   border: 1px solid #ccc;
   border-radius: 0.3rem;
 }
